Phase 2 bunch D: particle system to 1.21 codecs and render API
Three separate 1.21 API migrations, all confined to the particle package.

1. ParticleOptions.Deserializer was removed in 1.20.5. Create's
   ICustomParticleData now requires MapCodec<T> getCodec(ParticleType<T>) and
   StreamCodec<? super RegistryFriendlyByteBuf, T> getStreamCodec(). Migrated
   SmokeParticleData, PuffSmokeParticleData and ChimneyPushParticleData:
   RecordCodecBuilder.create -> mapCodec, makeDeserializer -> makeStreamCodec
   built from StreamCodec.composite, and dropped writeToNetwork /
   writeToString / getDeserializer (all gone from ParticleOptions).

2. VertexConsumer: vertex/uv/color/uv2/endVertex ->
   addVertex/setUv/setColor/setLight, with endVertex removed entirely.

3. ParticleRenderType is now an interface whose begin(Tesselator,
   TextureManager) RETURNS the BufferBuilder; end(Tesselator) is gone.

   BEHAVIOUR NOTE: TRANSPARENT_SMOKE's teardown (depthMask(true),
   disableBlend, defaultBlendFunc) had no home to move to -- 1.21 has no
   per-type teardown hook and vanilla types set state in begin() only. Worth
   watching for smoke-related depth/blend artifacts during client testing.

Also: TextureAtlasSprite.getV takes a float in 1.21, so the two call sites in
CustomAnimatedTextureSheetParticle narrow at the call rather than changing
the existing double-precision arithmetic.

A first attempt stripped (double) casts with a blanket regex, which also hit
`age / (double) this.lifetime` -- silently turning it into integer division.
That compiles cleanly and would only have shown up as wrong particle
behaviour in game. Reverted and redone with a paren-aware rewriter that only
strips casts inside addVertex(...) argument lists.

particle package errors: 60 -> 0
common compile errors: 608 -> 548

+++ common/src/main/java/com/railwayteam/railways/content/smokestack/particles/CustomAnimatedTextureSheetParticle.java
@@ -47,13 +47,13 @@ public abstract class CustomAnimatedTextureSheetParticle extends TextureSheetPar
     protected float getV0() {
         int frames = (this.sprite.contents().height() * frameWidthFactor()) / (this.sprite.contents().width() * frameHeightFactor());
         int frameNumber = (int) (frames * getAnimationProgress());
-        return this.sprite.getV(16. * ((double) frameNumber / frames));
+        return this.sprite.getV((float) (16. * ((double) frameNumber / frames)));
     }
 
     @Override
     protected float getV1() {
         int frames = (this.sprite.contents().height() * frameWidthFactor()) / (this.sprite.contents().width() * frameHeightFactor());
         int frameNumber = (int) (frames * getAnimationProgress());
-        return this.sprite.getV(16. * ((double) frameNumber + 1) / frames);
+        return this.sprite.getV((float) (16. * ((double) frameNumber + 1) / frames));
     }
 }
